+++ Sources/Sentry/include/SentryANRTrackingIntegration.h
@@ -1,12 +1,11 @@
 #import "SentryANRTrackerV1.h"
 #import "SentryBaseIntegration.h"
-#import "SentrySwift.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
 static NSString *const SentryANRExceptionType = @"App Hanging";
 
-@interface SentryANRTrackingIntegration : SentryBaseIntegration <SentryANRTrackerDelegate>
+@interface SentryANRTrackingIntegration : SentryBaseIntegration
 
 - (void)pauseAppHangTracking;
 - (void)resumeAppHangTracking;