+++ CHANGELOG.md
@@ -5,6 +5,13 @@
 ### Fixes
 
 - Add auto-start for session tracker when SDK is started after didBecomeActive (#5121)
+### Improvements
+
+- Add `itemCount` to `SentryEnvelopeItemHeader` ([#5230](https://github.com/getsentry/sentry-cocoa/pull/5230))
+
+### Features
+
+- Apps can now manually show and hide the included feedback widget button (#5236)
 
 ## 8.50.2